+++ src/almutentopicalswnd.py
@@ -43,8 +43,8 @@ class AlmutenTopicalsWnd(commonwnd.CommonWnd):
 
 		self.fntMorinus = ImageFont.truetype(common.common.symbols, self.FONT_SIZE)
 		self.fntText = ImageFont.truetype(common.common.abc, self.FONT_SIZE)
-		self.fntLargeText = ImageFont.truetype(common.common.abc, 5*self.FONT_SIZE/4)
-		self.fntBigText = ImageFont.truetype(common.common.abc, 3*self.FONT_SIZE/2)
+		self.fntLargeText = ImageFont.truetype(common.common.abc, int(5*self.FONT_SIZE/4))
+		self.fntBigText = ImageFont.truetype(common.common.abc, int(3*self.FONT_SIZE/2))
 		self.clrs = (self.options.clrdomicil, self.options.clrexal, self.options.clrperegrin, self.options.clrcasus, self.options.clrexil)	
 		self.signs = common.common.Signs1
 		if not self.options.signs:
@@ -70,7 +70,7 @@ class AlmutenTopicalsWnd(commonwnd.CommonWnd):
 		if self.bw:
 			tableclr = (0,0,0)
 
-		img = Image.new('RGB', (self.WIDTH, self.HEIGHT), self.bkgclr)
+		img = Image.new('RGB', (int(self.WIDTH), int(self.HEIGHT)), self.bkgclr)
 		draw = ImageDraw.Draw(img)
 
 		BOR = commonwnd.CommonWnd.BORDER
@@ -222,12 +222,12 @@ class AlmutenTopicalsWnd(commonwnd.CommonWnd):
 		sign = d/chart.Chart.SIGN_DEG
 		pos = d%chart.Chart.SIGN_DEG
 		wsp,hsp = draw.textsize(' ', self.fntText)
-		wsg,hsg = draw.textsize(self.signs[sign], self.fntMorinus)
+		wsg,hsg = draw.textsize(self.signs[int(sign)], self.fntMorinus)
 		txt = (str(pos)).rjust(2)+self.deg_symbol+(str(m)).zfill(2)+"'"+(str(s)).zfill(2)+'"'
 		w,h = draw.textsize(txt, self.fntText)
 		offset = (self.LONGITUDE_CELL_WIDTH-(w+wsp+wsg))/2
 		draw.text((x+offset, y+(self.LINE_HEIGHT-h)/2), txt, fill=clr, font=self.fntText)
-		draw.text((x+offset+w+wsp, y+(self.LINE_HEIGHT-hsg)/2), self.signs[sign], fill=clr, font=self.fntMorinus)
+		draw.text((x+offset+w+wsp, y+(self.LINE_HEIGHT-hsg)/2), self.signs[int(sign)], fill=clr, font=self.fntMorinus)
